React: Rename components to PascalCase for refactored components and format
