+++ static/js/redux/ui/CreditTicker.tsx
