+++ static/js/redux/ui/DayCalendar.tsx
