+++ static/js/redux/ui/Semesterly.tsx
