+++ static/js/redux/ui/containers/day_calendar_container.jsx
@@ -13,7 +13,7 @@ GNU General Public License for more details.
 */
 
 import { connect } from "react-redux";
-import DayCalendar from "../day_calendar";
+import DayCalendar from "../DayCalendar";
 import { saveTimetable } from "../../actions/user_actions";
 import { handleCreateNewTimetable } from "../../actions/timetable_actions";
 import {
